Move the status-change funnel to appointment-status.js
updateAppointmentStatus moves out of appointment-data.js into its own
module above appointment-data and episodes, ending the require cycle
that appointment-data previously dodged with a lazy require of
episodes.js. The funnel now builds the status update through
updateAppointmentData rather than duplicating its array/temp-copy
bookkeeping. Callers import from the new module; behaviour unchanged
(verified: check-in, cancel, undo-cancel flow).
