+++ app/routes/appointments/medical-history.js
@@ -9,9 +9,9 @@ const {
 } = require('../../lib/utils/participants')
 const {
   getAppointment,
-  saveTempAppointmentToAppointment,
-  updateAppointmentStatus
+  saveTempAppointmentToAppointment
 } = require('../../lib/utils/appointment-data')
+const { updateAppointmentStatus } = require('../../lib/utils/appointment-status')
 const generateId = require('../../lib/utils/id-generator')
 const {
   getReturnUrl,
